backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-pptx==0.6.23
python-multipart==0.0.6
Pillow==10.1.0
pydantic==2.5.0
httpx>=0.25.0
lxml>=4.9.0

# Testing
pytest>=7.4.0
hypothesis>=6.92.0
